Weighted Graphs
Graphs now have the ability to be both weighted and directed.

The big difference is that now the file to read in can ONLY have 3
parameters a line. The vertex that is the source of the edge. The vertex
that the source's edge is directed to. And the weight of that edge.

To make a graph unweighted, just make all of the weight file parameters
1.